+++ pymarkdown/plugins/rule_md_019.py
@@ -5,7 +5,7 @@ mark on an atx heading.
 from typing import Optional, cast
 
 from pymarkdown.general.parser_helper import ParserHelper
-from pymarkdown.plugin_manager.plugin_details import PluginDetails
+from pymarkdown.plugin_manager.plugin_details import PluginDetailsV2
 from pymarkdown.plugin_manager.plugin_scan_context import PluginScanContext
 from pymarkdown.plugin_manager.rule_plugin import RulePlugin
 from pymarkdown.tokens.atx_heading_markdown_token import AtxHeadingMarkdownToken
@@ -23,18 +23,18 @@ class RuleMd019(RulePlugin):
         super().__init__()
         self.__atx_heading_token: Optional[MarkdownToken] = None
 
-    def get_details(self) -> PluginDetails:
+    def get_details(self) -> PluginDetailsV2:
         """
         Get the details for the plugin.
         """
-        return PluginDetails(
+        return PluginDetailsV2(
             plugin_name="no-multiple-space-atx",
             plugin_id="MD019",
             plugin_enabled_by_default=True,
             plugin_description="Multiple spaces are present after hash character on Atx Heading.",
             plugin_version="0.5.0",
-            plugin_interface_version=1,
             plugin_url="https://github.com/jackdewinter/pymarkdown/blob/main/docs/rules/rule_md019.md",
+            plugin_supports_fix=True,
         )
 
     def starting_new_file(self) -> None:
@@ -42,6 +42,21 @@ class RuleMd019(RulePlugin):
         Event that the a new file to be scanned is starting.
         """
         self.__atx_heading_token = None
+
+    def __report(
+        self, context: PluginScanContext, text_token: TextMarkdownToken
+    ) -> None:
+        assert self.__atx_heading_token is not None
+        if context.in_fix_mode:
+            self.register_fix_token_request(
+                context,
+                text_token,
+                "next_token",
+                "extracted_whitespace",
+                " ",
+            )
+        else:
+            self.report_next_token_error(context, self.__atx_heading_token)
 
     def next_token(self, context: PluginScanContext, token: MarkdownToken) -> None:
         """
@@ -59,4 +74,4 @@ class RuleMd019(RulePlugin):
                 text_token.extracted_whitespace
             )
             if self.__atx_heading_token and len(resolved_extracted_whitespace) > 1:
-                self.report_next_token_error(context, self.__atx_heading_token)
+                self.__report(context, text_token)